Move basic test tools in utils
